+++ lib/ventriloquist/cap/platforms/debian/erlang_install.rb
@@ -3,22 +3,25 @@ module VagrantPlugins
     module Cap
       module Debian
         module ErlangInstall
-          ERLANG_SOLUTIONS_PKG = "https://packages.erlang-solutions.com/erlang-solutions_1.0_all.deb"
+          ERLANG_SOLUTIONS_PKG = "http://packages.erlang-solutions.com/erlang-solutions_1.0_all.deb"
 
           def self.erlang_install(machine)
             machine.communicate.tap do |comm|
-              if ! comm.test('which erl > /dev/null')
-                machine.env.ui.info('Installing Erlang')
-
-                path = download_path(comm)
-                unless comm.test("test -f #{path}")
-                  machine.guest.capability(:download, ERLANG_SOLUTIONS_PKG, path)
-                end
-                comm.sudo("dpkg -i #{path}")
-
-                comm.sudo('apt-get update')
-                comm.sudo('apt-get -y install erlang')
+              if comm.test('which erl > /dev/null')
+                machine.env.ui.info("Skipping Erlang installation")
+                return
               end
+
+              machine.env.ui.info('Installing Erlang')
+
+              path = download_path(comm)
+              unless comm.test("test -f #{path}")
+                machine.guest.capability(:download, ERLANG_SOLUTIONS_PKG, path)
+              end
+              comm.sudo("dpkg -i #{path}")
+
+              comm.sudo('apt-get update')
+              comm.sudo('apt-get -y install erlang')
             end
           end