+++ lib/ventriloquist/platforms/erlang.rb
@@ -3,6 +3,9 @@ module VagrantPlugins
     module Platforms
       class Erlang < Platform
         def provision(machine)
+          if ! @config[:versions].empty?
+            machine.env.ui.warn('Multiple versions of erlang were specified but the latest one will be installed')
+          end
           machine.guest.capability(:erlang_install)
         end
       end