+++ lib/ventriloquist/platforms/python.rb
@@ -3,10 +3,17 @@ module VagrantPlugins
     module Platforms
       class Python < Platform
         def provision(machine)
-          @config[:version] = '3.3.2' if @config[:version] == 'latest'
+          if @config[:versions].empty?
+            machine.env.ui.warn('No python version was specified and only pyenv will be installed')
+          end
+
           machine.guest.tap do |guest|
             guest.capability(:pyenv_install)
-            guest.capability(:pyenv_install_python, @config[:version])
+            # Reverse array so that the first version specified is installed last
+            # and gets set as the default
+            @config[:versions].reverse.each do |version|
+              guest.capability(:pyenv_install_python, version)
+            end
           end
         end
       end