+++ lib/ventriloquist/plugin.rb
@@ -1,7 +1,3 @@
-require "vagrant"
-
-Vagrant.require_plugin "vocker"
-
 require_relative "version"
 
 I18n.load_path << File.expand_path(File.dirname(__FILE__) + '/../../locales/en.yml')