+++ lib/ventriloquist/service.rb
@@ -17,11 +17,10 @@ module VagrantPlugins
         else
           machine.env.ui.info("Starting '#{@name}' service")
           machine.communicate.sudo("mkdir -p #{CONTAINER_IDS_PATH}")
-          # Reduce network latency, see https://groups.google.com/d/msg/docker-user/Z3zQuRawIsE/2AEkl30WpTQJ
-          # for more info
-          @config[:dns]     = '127.0.0.1'
-          @config[:name]    = @name
-          @config[:cidfile] = cidfile
+          @config[:cidfile]          = cidfile
+          @config[:daemonize]        = true
+          @config[:name]             = @name
+          @config[:auto_assign_name] = true
           @docker_client.run_container(@config)
         end
       end