+++ spec/unit/platforms_builder_spec.rb
@@ -5,12 +5,8 @@ require 'ventriloquist/platforms_builder'
 describe VagrantPlugins::Ventriloquist::PlatformsBuilder do
   Platform = VagrantPlugins::Ventriloquist::Platform
 
-  verify_contract(:platforms_builder)
-
-  fake(:docker_client)
-
-  let(:cfg)               { {version: '13.0'} }
-  let(:platforms_configs) { [{my_lang: cfg}, :your_lang, ['my_plat:version']] }
+  let(:cfg)               { {versions: '13.0'} }
+  let(:platforms_configs) { [{my_lang: cfg}, 'your_lang-1.2', ['my_plat-version']] }
   let(:custom_mapping)    { {'my_lang' => my_lang_class, 'your_lang' => your_lang_class, 'my_plat' => my_plat_class} }
 
   let(:my_lang_class) do
@@ -30,18 +26,18 @@ describe VagrantPlugins::Ventriloquist::PlatformsBuilder do
   let(:my_plat)   { platforms[2] }
 
   it 'builds a list of platform objects' do
-    expect(platforms).to have(3).items
+    expect(platforms.size).to eq(3)
     expect(my_lang).to   be_a(my_lang_class)
     expect(your_lang).to be_a(your_lang_class)
     expect(my_plat).to   be_a(my_plat_class)
   end
 
   it 'extracts version from platform name' do
-    expect(my_plat.config[:version]).to eq('version')
+    expect(my_plat.config[:versions]).to eq(['version'])
   end
 
   it 'defaults configured version to latest' do
-    expect(your_lang.config[:version]).to eq('latest')
+    expect(your_lang.config[:versions]).to eq(['1.2'])
   end
 
   it 'configures services using defined configs' do